backend/pyproject.toml
[project]
name = "ai-chatbot-rag"
version = "0.1.0"
description = "AI Chatbot with RAG - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "openai>=1.30.0",
    "tiktoken>=0.7.0",
    "python-multipart>=0.0.9",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "pgvector>=0.3.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
